Segregate methods from DriverBase
