+++ src/TripodStatFactory.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Tripod;
 
-use Tripod\Mongo\NoStat;
-
 class TripodStatFactory
 {
     /**
